interacting/todo/cmd/todo/main.go
package main

import (
	"flag"
	"fmt"
	"os"
	"small_console_applications_go/interacting/todo"
)

// Hardcoding file name for now
const todoFileName = ".todo.json"

func main() {
	// Parsing command line flags
	task := flag.String("task", "", "Task to be included in the todo list")
	list := flag.Bool("list", false, "List all tasks")
	complete := flag.Int("complete", 0, "Item to be completed")
	flag.Parse()
	l := &todo.List{}
	if err := l.Get(todoFileName); err != nil {
		fmt.Fprintln(os.Stderr, err)
		os.Exit(1)
	}
	// Divide what to do based on the number of arguments provided
	switch {
	case *list:
		fmt.Print(l)

	case *complete > 0:
		// Complete a task from the list
		if err := l.Complete(*complete); err != nil {
			fmt.Fprintln(os.Stderr, err)
			os.Exit(1)
		}
		// Save the list
		if err := l.Save(todoFileName); err != nil {
			fmt.Fprintln(os.Stderr, err)
			os.Exit(1)
		}
	case *task != "":
		// Add a task to the list
		l.Add(*task)
		// Save the list
		if err := l.Save(todoFileName); err != nil {
			fmt.Fprintln(os.Stderr, err)
			os.Exit(1)
		}
	default:
		// Invalid flag provided
		fmt.Println("Invalid option")
		os.Exit(1)
	}
}
